Move `GetTreestate` to the `trees` module
